Fixed deleting a category

+++ Services/Menu/MenuDeleteService.php
@@ -1,10 +1,12 @@
 <?php
 
 // Objects
+require_once __DIR__ . '/../../Objects/Menu/BilingualMenuCategory.php';
 
 // Repos
 require_once __DIR__ . '/../../Repository/Menu/MenuDeleteRepository.php';
 require_once __DIR__ . '/../../Repository/Menu/MenuReadRepository.php';
+require_once __DIR__ . '/../../Repository/Menu/MenuCategoryRepository.php';
 
 // Services
 require_once __DIR__ . '/../Shared/SortingService.php';
@@ -24,17 +26,20 @@ class MenuDeleteService
      * @var MenuDeleteRepository
      */
     private $deleteRepo;
-
     /**
      * @var MenuReadRepository
      */
     private $readRepo;
+    /**
+     * @var MenuCategoryRepository
+     */
+    private $categoryRepo;
 
     // Variables
     /**
-     * @var Message
+     * @var BilingualMenuCategory
      */
-    private $message;
+    private $category;
 
     public function __construct()
     {
@@ -45,40 +50,91 @@ class MenuDeleteService
         // Repos
         $this->deleteRepo = new MenuDeleteRepository();
         $this->readRepo = new MenuReadRepository();
+        $this->categoryRepo = new MenuCategoryRepository();
 
-        //Variables
-        $this->message = new Message();
     }
 
+    /**
+     * @todo: Check if category exists.
+     *      Create menu category with result
+     *      Get all (if any) items
+     *      Then delete item descriptions (this works already)
+     *      Then delete items (this doesn't work)
+     *      Then delete category
+     */
+
+    /**
+     * @param $categoryId
+     * @return bool
+     */
     public function initializeCategoryDelete($categoryId)
     {
-
-        $data = $this->readRepo->getAllItemsByCategory($categoryId);
-        if ($data !== false) {
-            $this->deleteCategory($data);
+        $check = true;
+        if ($this->checkCategoryExists($categoryId)) {
+            $data = $this->readRepo->getAllItemsByCategory($categoryId);
+            if ($data !== false) {
+                $this->deleteCategory($data);
+            } else {
+                if (!$this->deleteRepo->deleteCategory($categoryId)) {
+                    $check = false;
+                }
+            }
         } else {
+            $check = false;
         }
 
-        return $this->message;
+        return $check;
+    }
+
+    private function sortMenuResults($rawResults)
+    {
+        foreach ($rawResults as $result) {
+            if (!isset($sortedResults)) {
+                if ($result['catId'] !== null) {
+                    $sortedResults = new BilingualMenuCategory($result['catEnName'],
+                        $result['catVnName'], $result['catType'], $result['pagePosition'],
+                        $result['catId'], $result['createdAt'], $result['editedAt']);
+                }
+            }
+
+            //if (isset($result['itemId']))
+        }
+
+        return $sortedResults;
+    }
+
+    public function checkCategoryExists(int $categoryId)
+    {
+        $check = false;
+        $categories = $this->categoryRepo->getCategories();
+
+        foreach ($categories as $category) {
+            if ($category->id === $categoryId) {
+                $this->category = $category;
+                $check = true;
+                break;
+            }
+        }
+
+        return $check;
     }
 
     public function initializeItemDelete(int $itemId)
     {
-
+        $check = true;
         $data = $this->readRepo->getItemById($itemId);
 
         if ($data !== false) {
-            if ($this->deleteItem($data)) {
+            if (!$this->deleteItem($data)) {
+                $check = false;
             }
         } else {
+            $check = false;
         }
 
-        return $this->message;
+        return $check;
     }
 
-    /**
-     * @param BilingualMenuCategory $category
-     */
     private function deleteCategory(BilingualMenuCategory $category)
     {
 
@@ -90,13 +146,13 @@ class MenuDeleteService
                     $check = false;
                 }
             }
-        } else {
+        }
+        if (!$this->deleteRepo->deleteCategory($category->id)) {
             $check = false;
         }
-        if ($check === true) {
-            if (!$this->deleteRepo->deleteCategory($category->id)) {
-            }
-        }
+
+        return $check;
+
     }
 
     /**
@@ -108,11 +164,17 @@ class MenuDeleteService
 
         $check = true;
 
-        if ($this->deleteDescription($item->enId) && $this->deleteDescription($item->vnId)) {
-            if (!$this->deleteRepo->deleteItem($item->itemId)) {
+        if ($item->enId !== null) {
+            if (!$this->deleteDescription($item->enId)) {
                 $check = false;
             }
-        } else {
+        }
+        if ($item->vnId !== null) {
+            if ($this->deleteDescription($item->vnId)) {
+                $check = false;
+            }
+        }
+        if (!$this->deleteRepo->deleteItem($item->itemId)) {
             $check = false;
         }
